feat(Auth): added login, register pages, jwt auth middleware
-  create login and register pages using react hook forms
- to sign a jwt token we need a secret key which we store in our .env file. use dotenv library to load env variables in .env into process.env.
- then when we login we sign a jwt token. the jwt sign method takes a payload and a secret key. we await it and send it along with the response to login. the jet token will take the format headerbase64.payloadbase64.signature
- we add a authenticateToken middleware which will verify the jwt token. im using postman, i have added the authorization header and added the token to it.
its recommended to access the authorisation property in headers prop of req through bracket notation for consistency. headers are usually accessed this way. now we have the token. now we pass in the token and the secret key to jwt.verify. it takes the header and payload from the token and the secret key to calculate the token again and checks whether they match if they match we call next which calls our controller. we use async await syntax to do this asynchronously.

#8